Inclusao do algoritmo A* com heuristica simples
