+++ nbproject/private/private.xml
@@ -3,9 +3,10 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/C:/sistemas/8puzzle/src/view/TelaInicial.java</file>
-            <file>file:/C:/sistemas/8puzzle/src/view/Tabuleiro.java</file>
-            <file>file:/C:/sistemas/8puzzle/src/control/Jogo2.java</file>
+            <file>file:/home/daniel/CodigosFontes/Java/8puzzle/src/view/Tabuleiro.java</file>
+            <file>file:/home/daniel/CodigosFontes/Java/8puzzle/src/pkg8puzzle/Main.java</file>
+            <file>file:/home/daniel/CodigosFontes/Java/8puzzle/src/view/TelaInicial.java</file>
+            <file>file:/home/daniel/CodigosFontes/Java/8puzzle/src/control/Jogo2.java</file>
         </group>
     </open-files>
 </project-private>
