--- src/helpers/Helper.java
@@ -1,24 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package helpers;
-
-/**
- *
- * @author Elizeu-pc
- */
-public class Helper {
-
-    public static <T> int indexOf(T needle, T[] haystack) {
-        for (int i = 0; i < haystack.length; i++) {
-            if (haystack[i] != null && haystack[i].equals(needle)
-                    || needle == null && haystack[i] == null) {
-                return i;
-            }
-        }
-
-        return -1;
-    }
-}
